backend/requirements.txt
Django==3.2.10
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-modeltranslation==0.18.2
django-rest-auth==0.9.5
django-taggit==3.0.0
django-templated-mail==1.1.1
django-translator==1.0.15
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2-binary==2.9.1
psycopg2==2.9.3
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
django-oauth-toolkit==2.1.0